+++ langchain_hana/__init__.py
@@ -1,7 +1,7 @@
 from importlib import metadata
 
-from langchain_sap_hana_cloud.embeddings import HanaInternalEmbeddings
-from langchain_sap_hana_cloud.vectorstores import HanaDB
+from langchain_hana.embeddings import HanaInternalEmbeddings
+from langchain_hana.vectorstores import HanaDB
 
 try:
     __version__ = metadata.version(__package__)
