+++ langchain_hana/utils.py
