+++ pyproject.toml
@@ -3,9 +3,9 @@ requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
-name = "langchain-sap-hana-cloud"
+name = "langchain-hana"
 version = "0.1.0"
-description = "An integration package connecting SAP Hana Cloud and LangChain"
+description = "An integration package connecting SAP HANA Cloud and LangChain"
 authors = []
 readme = "README.md"
 license = "Apache-2.0"
