+++ tests/integration_tests/test_vectorstores.py
@@ -7,8 +7,8 @@ from typing import Any, Dict, List
 import numpy as np
 import pytest
 
-from langchain_sap_hana_cloud.utils import DistanceStrategy
-from langchain_sap_hana_cloud.vectorstores import HanaDB
+from langchain_hana.utils import DistanceStrategy
+from langchain_hana.vectorstores import HanaDB
 from tests.integration_tests.fake_embeddings import ConsistentFakeEmbeddings
 from tests.integration_tests.fixtures.filtering_test_cases import (
     DOCUMENTS,
